+++ TesteCalculoSeguro.Application/Services/ISeguroService.cs
@@ -5,10 +5,10 @@ namespace TesteCalculoSeguro.Application.Services
     public interface ISeguroService
     {
         IEnumerable<Veiculo> ObterCarros();
-        double CalcularValorSeguro(double valorVeiculo);
-        double CalcularTaxaDeRisco(double valorVeiculo);
-        double CalcularPremioDeRisco(double taxaDeRisco, double valorVeiculo);
-        double CalcularPremioPuro(double premioDeRisco);
-        double CalcularPremioComercial(double premioPuro);
+        decimal CalcularValorSeguro(decimal valorVeiculo);
+        decimal CalcularTaxaDeRisco(decimal valorVeiculo);
+        decimal CalcularPremioDeRisco(decimal taxaDeRisco, decimal valorVeiculo);
+        decimal CalcularPremioPuro(decimal premioDeRisco);
+        decimal CalcularPremioComercial(decimal premioPuro);
     }
 }
